Update Dockerfile
fix for missing "cffi.h"  header file during the wheel build of "cryptography" that is a dependency of jwt from the requirements.txt file.

I tried to add the python3-libffi it to the requirements.txt file, but this didn't fix the issue for me.  I  had to install the Debian package "python3-cffi" and "libffi-dev" for the Docker container to sucessfully build.